requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
asyncpg==0.29.0
bcrypt==4.1.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.0
fastapi==0.110.1
google-api-core==2.19.0
google-api-python-client==2.130.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
Mako==1.3.3
MarkupSafe==2.1.5
oauth2client==4.1.3
oauthlib==3.2.2
packaging==24.0
passlib==1.7.4
pluggy==1.5.0
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.8.0
pymongo==4.6.3
pyparsing==3.1.2
pytest==8.2.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.32.2
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.29.0
watchfiles==0.21.0
websockets==12.0
